third_party/blink/web_tests/platform/mac-mac13-arm64/fast/webgl/tex-sub-image-cube-maps-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'createBuffer')
Checks texSubImage2D with cube map textures

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



FAIL Unable to fetch WebGL rendering context for Canvas
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
